## 2025/10/25 user interface_refinement-01 - rudsonalves

+++ lib/data/repositories/checkpoint_data_repository.dart
@@ -1,6 +1,8 @@
+import 'dart:developer';
+
 import 'package:result_dart/result_dart.dart';
 
-import '/core/extensions/result.dart';
+import '/core/extensions/string_extension.dart';
 import '/core/services/secure_storage/local_secure_storage.dart';
 import '/domain/entities/checkpoint/checkpoint_data.dart';
 import '/domain/entities/checkpoint/checkpoint_values/base_checkpoint_values.dart';
@@ -11,82 +13,105 @@ class CheckpointDataRepositoryImpl implements CheckpointDataRepository {
 
   CheckpointDataRepositoryImpl(this._secureStorage);
 
-  static const String _storageKey = 'checkpoint_data';
+  static const String _storageKey = 'checkpoint';
 
   CheckpointData _checkpointData = CheckpointData.empty();
-  bool _isEdited = false;
 
   @override
   CheckpointData get checkpointData => _checkpointData;
 
   @override
   AsyncResult<CheckpointData> getCheckpointData() async {
-    return await _secureStorage
-        .read(_storageKey)
-        .flatMap(
-          (source) => toResult(() => CheckpointData.fromJson(source)).fold(
-            (checkpoint) {
-              _checkpointData = checkpoint;
-              return Success(checkpoint);
-            },
-            (err) {
-              _checkpointData = CheckpointData.empty();
-              return Failure(Exception('Erro ao carregar dados: $err'));
-            },
-          ),
-        );
-    // TODO: Rudson - load from API if no checkpoint data found in local storage
-    //       Eventually search the API first!!!
+    final result = await _secureStorage.read(_storageKey);
+
+    if (result is Failure) {
+      _checkpointData = CheckpointData.empty();
+      return Failure(result.exceptionOrNull() ?? Exception('Internal Error'));
+    }
+
+    final source = (result as Success).getOrNull() as String;
+
+    final checkpointData = CheckpointData.fromJson(source);
+
+    final sections = <CheckpointStage, CheckpointSectionData>{};
+    for (final stage in [
+      CheckpointStage.createPersonalAccount,
+      CheckpointStage.createBusinessAccount,
+      CheckpointStage.registerBusinessPartners,
+    ]) {
+      final key = 'checkpoint_${stage.stageName}';
+      final jsonStr = (await _secureStorage.read(key)).getOrNull();
+
+      if (jsonStr != null && jsonStr.isNotEmpty) {
+        try {
+          sections[stage] = CheckpointSectionData.fromJson(
+            stage: stage,
+            json: jsonStr,
+          );
+        } catch (err) {
+          // Ignore
+        }
+      }
+    }
+    _checkpointData = checkpointData.copyWith(sections: sections);
+
+    return Success(_checkpointData);
   }
 
   @override
   AsyncResult<Unit> createCheckpointSection<T extends BaseCheckpointValues>(
     T values,
   ) async {
-    final CheckpointStage stage = values.stage;
+    final stage = values.stage;
 
     if (_checkpointData.hasSectionData(stage)) {
       return Success(unit);
     }
 
-    final updatedSections = Map<CheckpointStage, CheckpointSectionData>.from(
+    final sections = Map<CheckpointStage, CheckpointSectionData>.from(
       _checkpointData.sections,
     );
-    updatedSections[stage] = CheckpointSection<T>(values: values);
-    final newCheckpointData = _checkpointData.copyWith(
-      sections: updatedSections,
+    sections[stage] = CheckpointSection<T>(values: values);
+
+    _checkpointData = _checkpointData.copyWith(
       currentStage: stage,
+      sections: sections,
     );
-    return await saveCheckpointData(newCheckpointData);
+
+    return await saveCheckpointData();
   }
 
   @override
-  AsyncResult<Unit> saveCheckpointData(CheckpointData checkpointData) async {
-    try {
-      final jsonString = checkpointData.toJson();
-      final writeResult = await _secureStorage.write(_storageKey, jsonString);
+  AsyncResult<Unit> saveCheckpointData() async {
+    if (!_checkpointData.hasAnyDirtyFields) return Success(unit);
 
-      return writeResult.fold(
-        (success) {
-          _isEdited = _isEdited || _checkpointData != checkpointData;
-          _checkpointData = checkpointData;
+    return await _secureStorage
+        .write(_storageKey, _checkpointData.toJson())
+        .flatMap((_) async {
+          for (final entry in _checkpointData.sections.entries) {
+            final sectionData = entry.value;
+
+            if (sectionData is CheckpointSection &&
+                sectionData.values.isDirty) {
+              final key = '${_storageKey}_${entry.key.stageName}';
+              await _secureStorage.write(key, sectionData.toJson());
+            }
+          }
+
           return Success(unit);
-        },
-        (error) {
-          return Failure(Exception('Erro ao salvar dados: $error'));
-        },
-      );
-    } catch (err) {
-      return Failure(Exception('Erro ao salvar dados: $err'));
-    }
+        });
   }
 
   @override
   AsyncResult<Unit> syncWithApi() async {
     try {
-      if (!_isEdited) return Success(unit);
-      // TODO: Implementar sincronização com API quando necessário
-      _isEdited = false;
+      if (!_checkpointData.hasAnyDirtyFields) return Success(unit);
+      await Future.delayed(const Duration(milliseconds: 800));
+
+      log('----------Not implemented----------');
+      log(_checkpointData.toJson());
+      log('-----------------------------------');
+
       return const Success(unit);
     } catch (err) {
       return Failure(Exception('Erro na sincronização: $err'));
@@ -99,37 +124,17 @@ class CheckpointDataRepositoryImpl implements CheckpointDataRepository {
   }
 
   @override
-  AsyncResult<CheckpointData> updateSection({
-    required CheckpointStage stage,
-    required CheckpointSectionData sectionData,
-    CheckpointStage? nextStage,
-  }) async {
-    try {
-      final updatedSections = Map<CheckpointStage, CheckpointSectionData>.from(
-        _checkpointData.sections,
-      );
-      updatedSections[stage] = sectionData;
-      final updatedCheckpoint = _checkpointData.copyWith(
-        sections: updatedSections,
-        currentStage: nextStage ?? stage,
-      );
-      final saveResult = await saveCheckpointData(updatedCheckpoint);
-      return saveResult.fold(
-        (failure) => Failure(Exception('Falha ao salvar seção: $failure')),
-        (_) => Success(updatedCheckpoint),
-      );
-    } catch (err) {
-      return Failure(Exception('Erro ao atualizar seção: $err'));
-    }
-  }
-
-  @override
-  AsyncResult<Unit> clearCheckpointData() async {
+  AsyncResult<Unit> clearCheckpointData({String? cpf}) async {
     try {
       _checkpointData = CheckpointData.empty();
-      // TODO: Rudson - update in the api too
-      return await _secureStorage.delete(_storageKey);
+      if (cpf == null) return await _secureStorage.delete(_storageKey);
+
+      _checkpointData = CheckpointData.newPerson();
+      _checkpointData.personalAccountValues!.cpf = cpf.onlyNumbers();
+
+      return saveCheckpointData();
     } catch (err) {
+      log(err.toString());
       return Failure(Exception('Erro ao limpar dados: $err'));
     }
   }
